Store directory geometry
Use Berkelee DB to store the x, y, h, w for each directory. Use the hash
storage type.

These data are persisted in `~/.argonautinfo`, or in `/.argonautinfo` if
the user has no `$HOME`.

We record the geometry after the user has moved the window -
specifically, after the user has reconfigured the window.

Guidance for OpenBSD added to the README.