Become a drop target
This is the drop in drag-and-drop.

The icon view is a drop destination for text and URI targets. Dirctories
are highlighted as the mouse moves over them; otherwise, the icon view
itself is highlighted. This reflects what happens when the drop is
released: the files are moved/copied to either the directory the mouse
was on, or to the window's directory.

If the source and target are on different mount points, copy the file;
otherwise, move the file. Works for directories and files. Does not
work for multiple sources.

Unrelated: the compatibility files were always being included,
overriding OS functions. Fix the whole compat mechanism so that it works
correctly.