Add a desktop entry
The desktop entry specification is used by panels to generate a list of
programs to present to the user. This includes GNOME2, MATE, GNOME3,
fbpanel, KDE, and more.

We include a desktop entry specification with argonaut so that it will
show in people's panels.

Hardcode the path to argonaut in the desktop entry to simplify developer
installation.

This commit also adds a list of other Free Desktop standards that are
relevant to this project.