Show a desktop file manager
This shows `$HOME` as the desktop. If `$HOME` is undefined, show `/`.

The icons are shows horizontally. This is the easiest thing to do with a
`GtkIconView`. Changing it later is left as an exercise for the reader.

The desktop color can be configured by your GTK+ 3 theme by selecting
`#argonaut-desktop`. See the example in the manpage for
more.

Single middle-clicking on an icon opens it in an argonaut(1) window. It
finds the `argonaut` binary via the `@BINDIR` autotools variable. If the
binary cannot be found, print an error to `stderr`.