Dragging files and directories
Completing the other half of DnD.

After the drag is complete, repopulate the window. While here,
repopulate the window after a drop is complete, too.

This only supports text/uri-list, which is fine in my 10 minutes of
practice.

This commit also includes some refactorings including separate source
files for drag and for drop, more header files, and more utility
functions.